// 05VUE/sc-nestjs/src/pojo/interface/payload.interface.ts
/*
 * @Author: Nero
 * @Github: https://github.com/nero90011
 * @E-mail: 601412402@qq.com
 * @Company: myself
 * @Date: 2020-04-17 16:32:42
 * @LastEditTime: 2020-04-23 17:07:23
 * @LastEditors: Do not edit
 * @Description: 
 */
export interface Payload{
    username?:string
    sub?:number
    role?:number
    img?:string
    locations:string
}